myproject/requirements.txt
Django==5.2.6
djangorestframework
django-cors-headers
psycopg2-binary
